pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "resume-agent"
version = "0.1.0"
description = "一个基于 AI 的简历一键定制工具，支持结构化输出与 STAR 法则重写。"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "David", email = "david@example.com"}
]
keywords = ["AI", "Resume", "LLM", "Agent", "Pydantic"]
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-ai>=0.0.14",
    "python-dotenv>=1.0.0",
    "jinja2>=3.0.0",
    "playwright>=1.40.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.urls]
"Homepage" = "https://github.com/bowen-duan/resume-agent"
"Bug Tracker" = "https://github.com/bowen-duan/resume-agent/issues"

[project.scripts]
resume-run = "resume_agent.main:main"

[tool.setuptools.packages.find]
where = ["src"]